Pendant Key Chalenge - Code/requirements.txt
numpy
opencv-python
Pillow
moviepy
mediapipe
torch
torchvision
segment-anything